重净化量为负数问题，自动重置滤芯

+++ OznerLibrary/AirPurifier/MXChip/MxChipAirPurifierSensor.m
@@ -17,7 +17,7 @@
     }
     return self;
 }
--(short)getInt:(int)propertyId
+-(int)getInt:(int)propertyId
 {
     @synchronized(propertys) {
         NSData* data=[propertys objectForKey:[NSString stringWithFormat:@"%d",propertyId]];
